backend/requirements-lambda.txt
fastapi==0.115.6
mangum==0.19.0
motor==3.6.0
beanie==1.27.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
httpx==0.28.1
